minor move around
