--- simapp/params.go
@@ -1,27 +1,0 @@
-package simapp
-
-// Simulation parameter constants
-const (
-	StakePerAccount                                    = "stake_per_account"
-	InitiallyBondedValidators                          = "initially_bonded_validators"
-	OpWeightDeductFee                                  = "op_weight_deduct_fee"
-	OpWeightMsgSend                                    = "op_weight_msg_send"
-	OpWeightSingleInputMsgMultiSend                    = "op_weight_single_input_msg_multisend"
-	OpWeightMsgSetWithdrawAddress                      = "op_weight_msg_set_withdraw_address"
-	OpWeightMsgWithdrawDelegationReward                = "op_weight_msg_withdraw_delegation_reward"
-	OpWeightMsgWithdrawValidatorCommission             = "op_weight_msg_withdraw_validator_commission"
-	OpWeightSubmitVotingSlashingTextProposal           = "op_weight_submit_voting_slashing_text_proposal"
-	OpWeightSubmitVotingSlashingCommunitySpendProposal = "op_weight_submit_voting_slashing_community_spend_proposal"
-	OpWeightSubmitVotingSlashingParamChangeProposal    = "op_weight_submit_voting_slashing_param_change_proposal"
-	OpWeightMsgDeposit                                 = "op_weight_msg_deposit"
-	OpWeightMsgCreateValidator                         = "op_weight_msg_create_validator"
-	OpWeightMsgEditValidator                           = "op_weight_msg_edit_validator"
-	OpWeightMsgDelegate                                = "op_weight_msg_delegate"
-	OpWeightMsgUndelegate                              = "op_weight_msg_undelegate"
-	OpWeightMsgBeginRedelegate                         = "op_weight_msg_begin_redelegate"
-	OpWeightMsgUnjail                                  = "op_weight_msg_unjail"
-	OpWeightMsgTransferNFT                             = "op_weight_msg_transfer_nft"
-	OpWeightMsgEditNFTMetadata                         = "op_weight_msg_edit_nft_metadata"
-	OpWeightMsgMintNFT                                 = "op_weight_msg_mint_nft"
-	OpWeightMsgBurnNFT                                 = "op_weight_msg_burn_nft"
-)
