+++ x/nft/app/genesis.go
@@ -2,6 +2,8 @@ package simapp
 
 import (
 	"encoding/json"
+
+	"github.com/cosmos/cosmos-sdk/types/module"
 )
 
 // The genesis state of the blockchain is represented here as a map of raw json
@@ -14,6 +16,6 @@ import (
 type GenesisState map[string]json.RawMessage
 
 // NewDefaultGenesisState generates the default state for the application.
-func NewDefaultGenesisState() GenesisState {
+func NewDefaultGenesisState(mb module.BasicManager) GenesisState {
 	return ModuleBasics.DefaultGenesis()
 }
